Use time manager interface

+++ Assets/Scripts/Strategy/Transitions/SceneSharing.cs
@@ -2,6 +2,6 @@
 {
     static class SceneSharing
     {
-        public static int timeStep;
+        public static ulong timeStep;
     }
 }
